+++ samples/expo/app/(tabs)/index.tsx
@@ -6,6 +6,8 @@ import { Text, View } from '@/components/Themed';
 import { SENTRY_INTERNAL_DSN } from '@/utils/dsn';
 import { HttpClient } from '@sentry/integrations';
 import { setScopeProperties } from '@/utils/setScopeProperties';
+import { timestampInSeconds } from '@sentry/utils';
+import React from 'react';
 
 const isRunningInExpoGo = Constants.appOwnership === 'expo'
 
@@ -38,6 +40,7 @@ Sentry.init({
         // default: [/.*/]
         failedRequestTargets: [/.*/],
       }),
+      Sentry.metrics.metricsAggregatorIntegration(),
     );
     return integrations.filter(i => i.name !== 'Dedupe');
   },
@@ -66,6 +69,25 @@ Sentry.init({
 });
 
 export default function TabOneScreen() {
+  const [componentMountStartTimestamp] = React.useState<number>(() => {
+    return timestampInSeconds();
+  });
+
+  React.useEffect(() => {
+    if (componentMountStartTimestamp) {
+      // Distributions help you get the most insights from your data by allowing you to obtain aggregations such as p90, min, max, and avg.
+      Sentry.metrics.distribution(
+        'tab_one_mount_time',
+        timestampInSeconds() - componentMountStartTimestamp,
+        {
+          unit: "seconds",
+        },
+      );
+    }
+    // We only want this to run once.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
   return (
     <View style={styles.container}>
       <Text>Welcome to Sentry Expo Sample App!</Text>
@@ -78,6 +100,7 @@ export default function TabOneScreen() {
       <Button
         title="Capture exception"
         onPress={() => {
+          Sentry.metrics.increment('tab_one.capture_exception_button_press', 1);
           Sentry.captureException(new Error('Captured exception'));
         }}
       />