+++ samples/react-native/src/Screens/TrackerScreen.tsx
@@ -40,6 +40,13 @@ const TrackerScreen = () => {
       });
   };
 
+  const onRefreshButtonPress = () => {
+    Sentry.metrics.increment('tracker_screen.refresh_button_press', 1, {
+      tags: { graph: 'none', public_data: true },
+    });
+    loadData();
+  };
+
   React.useEffect(() => {
     console.log('transitionEnd TrackerScreen mounted', timestampInSeconds() * 1000);
     loadData();
@@ -73,7 +80,11 @@ const TrackerScreen = () => {
           <ActivityIndicator size="small" color="#F6F6F8" />
         )}
       </View>
-      <Button sentry-label="refresh" title="Refresh" onPress={loadData} />
+      <Button
+        sentry-label="refresh"
+        title="Refresh"
+        onPress={onRefreshButtonPress}
+      />
     </View>
   );
 };